feat: render plan approval content as rich markdown
Extract shared markdown config (plugins, sanitization, preprocessing)
from ChatPanel into utils/markdown.ts and reuse it in PlanApprovalCard
to render plan files with full markdown support instead of plain text.
