+++ src/ui/src/components/chat/PlanApprovalCard.module.css
@@ -64,15 +64,90 @@
   background: var(--hover-bg-subtle);
   border: 1px solid var(--divider);
   border-radius: 8px;
-  padding: 14px;
+  padding: 14px 18px;
   margin-bottom: 14px;
-  max-height: 300px;
+  max-height: 400px;
   overflow-y: auto;
   font-size: 13px;
   line-height: 1.6;
   color: var(--text-primary);
-  white-space: pre-wrap;
+}
+
+/* Prose styles for rendered markdown inside the plan card */
+.planContent h1,
+.planContent h2,
+.planContent h3,
+.planContent h4 {
+  margin: 1em 0 0.5em;
+  color: var(--text-primary);
+  font-weight: 600;
+}
+
+.planContent h1 { font-size: 1.3em; }
+.planContent h2 { font-size: 1.15em; }
+.planContent h3 { font-size: 1.05em; }
+
+.planContent h1:first-child,
+.planContent h2:first-child,
+.planContent h3:first-child {
+  margin-top: 0;
+}
+
+.planContent p {
+  margin: 0.6em 0;
+}
+
+.planContent ul,
+.planContent ol {
+  margin: 0.5em 0;
+  padding-left: 1.5em;
+}
+
+.planContent li {
+  margin: 0.25em 0;
+}
+
+.planContent code {
   font-family: var(--font-mono, monospace);
+  font-size: 0.9em;
+  background: rgba(255, 255, 255, 0.06);
+  padding: 0.15em 0.4em;
+  border-radius: 4px;
+}
+
+.planContent pre {
+  margin: 0.6em 0;
+  padding: 10px 12px;
+  background: rgba(0, 0, 0, 0.25);
+  border-radius: 6px;
+  overflow-x: auto;
+}
+
+.planContent pre code {
+  background: none;
+  padding: 0;
+  font-size: 12px;
+  line-height: 1.5;
+}
+
+.planContent hr {
+  border: none;
+  border-top: 1px solid var(--divider);
+  margin: 1em 0;
+}
+
+.planContent strong {
+  color: var(--text-primary);
+  font-weight: 600;
+}
+
+.planContent a {
+  color: var(--accent-primary);
+  text-decoration: none;
+}
+
+.planContent a:hover {
+  text-decoration: underline;
 }
 
 .permissions {
